move a function from plotting.py to database.py
